Add minio service classes and update dependenciesThese code changes appear to be from the Herodotus OAuth2 and Captcha modules. Here are the key updates:For the OAuth2 module:1. Refactored imports to use static imports where appropriate.2. Updated error codes to use the constants from OAuth2ErrorCodes.3. Replaced some string literals with constants from OidcParameterNames.4. Improved code readability in some areas, such as using the new Java5+ enhanced for loop.5. Updated the way tokens are generated by using OAuth2TokenContext instead of DefaultOAuth2TokenContext.Builder.For the Captcha module:1. Improved code readability, such as using more specific variable types and removing unnecessary type casts.2. Updated some method calls to use newer Java methods (e.g., Arrays.asList() instead of explicit type).3. Refactored some methods to make them more concise and easier to understand.4. Updated the way some calculations are done, such as using (float)1 / length instead of1.0f / length for better precision.Overall, these changes improve code readability, maintainability, and consistency across the modules. They also make better use of newer Java features and follow best practices more closely.

+++ engine-core/core-definition/src/main/java/com/pigx/engine/core/definition/domain/view/datatables/DataTableUtils.java
@@ -2,8 +2,9 @@ package com.pigx.engine.core.definition.domain.view.datatables;
 
 import java.util.List;
 
-/* loaded from: core-definition-3.5.7.0.jar:cn/herodotus/engine/core/definition/domain/view/datatables/DataTableUtils.class */
+
 public class DataTableUtils {
+
     public static final String ECHO = "sEcho";
     public static final String DISPLAY_START = "iDisplayStart";
     public static final String DISPLAY_LENGTH = "iDisplayLength";
@@ -11,6 +12,7 @@ public class DataTableUtils {
     public static final String DATA = "data";
 
     public static DataTableResult parseDataTableParameter(List<DataTableParameter> params) {
+
         String sEcho = null;
         String jsonString = null;
         int iDisplayStart = 0;
@@ -20,15 +22,16 @@ public class DataTableUtils {
                 sEcho = param.getValue().toString();
             }
             if (param.getName().equals(DISPLAY_START)) {
-                iDisplayStart = ((Integer) param.getValue()).intValue();
+                iDisplayStart = (int) param.getValue();
             }
             if (param.getName().equals(DISPLAY_LENGTH)) {
-                iDisplayLength = ((Integer) param.getValue()).intValue();
+                iDisplayLength = (int) param.getValue();
             }
             if (param.getName().equals(QUERY_JSON)) {
                 jsonString = param.getValue().toString();
             }
         }
+
         return new DataTableResult(sEcho, iDisplayStart, iDisplayLength, jsonString);
     }
 }
